requirements.txt
BzrTools==2.6.0
Flask==0.12.2
Flask-Login==0.4.0
Flask-OpenID==1.2.5
Jinja2==2.9.6
MarkupSafe==1.0
MySQL-python==1.2.5
Pillow==2.3.0
SQLAlchemy==1.1.10
SecretStorage==2.0.0
Werkzeug==0.12.2
argparse==1.2.1
boto3==1.4.4
botocore==1.5.65
bzr==2.7.0dev1
certifi==2017.4.17
chardet==3.0.4
click==6.7
colorama==0.2.5
configobj==4.7.2
decorator==3.4.0
docutils==0.13.1
duplicity==0.6.23
ecdsa==0.13
envs==1.1.1
funcsigs==1.0.2
future==0.16.0
futures==3.1.1
html5lib==0.999
httplib2==0.8
idna==2.5
ipython==1.2.1
itsdangerous==0.24
jmespath==0.9.3
keyring==3.5
launchpadlib==1.10.2
lazr.restfulclient==0.13.3
lazr.uri==1.0.3
lockfile==0.8
lpthw.web==1.1
matplotlib==1.3.1
mercurial==2.8.2
mock==2.0.0
numpy==1.8.2
oauth==1.0.1
panda==0.3.1
paramiko==1.10.1
pbr==3.0.1
pexpect==3.1
pycrypto==2.6.1
pygobject==3.12.0
pygpgme==0.3
pyparsing==2.0.1
python-apt==0.9.3.5ubuntu2
python-dateutil==2.6.0
python-jose==1.3.2
python-openid==2.2.5
pytz==2012c
requests==2.17.3
s3transfer==0.1.10
scipy==0.13.3
simplegeneric==0.8.1
simplejson==3.3.1
six==1.10.0
stevedore==0.14.1
terminaltables==3.1.0
urllib3==1.21.1
virtualenv==1.11.4
virtualenv-clone==0.2.4
virtualenvwrapper==4.1.1
wadllib==1.3.2
warrant==0.2.0
wheel==0.24.0
wsgiref==0.1.2
zope.interface==4.0.5
